feat: 유저별 장바구니 리스트 조회 기능 구현
- 로그인한 유저의 고유 ID를 기반으로 장바구니 목록을 필터링하여 노출하도록 로직 수정
- [TODO] 장바구니 아이템 수량 및 가격 변동 시, DB 선반영 후 데이터를 리프레시하도록 UI 로직 수정 예정

+++ app/api/cart/all_delete_cart/route.ts
@@ -1,12 +1,15 @@
 import { connectDB } from "@/app/lib/database";
-import { NextResponse } from "next/server";
+import { NextRequest ,NextResponse } from "next/server";
 
 const dbName = process.env.DB_NAME;
 
-export async function DELETE() {
+export async function DELETE(request: NextRequest) {
+    const searchParams = request.nextUrl.searchParams;
+    const userId = searchParams.get("userId");
+
     try{
         const db = (await connectDB).db(dbName);
-        await db.collection('carts').deleteMany({});
+        await db.collection('carts').deleteMany({userId});
 
         return NextResponse.json({ success: true, message: "삭제가 완료되었습니다." }, {status: 200});
     }catch(err) {
